+++ frontend/src/domains/pickeat/utils/makeNickname.ts
